+++ src/Encoder/IntegerEncoder.php
@@ -33,8 +33,8 @@ class IntegerEncoder implements Encoder
             'decimal' => function ($value, $options) {
                 return $this->encodeDecimal($value, $options);
             },
-            'hexadecimal' => function ($value) {
-                return $this->encodeHexadecimal($value);
+            'hexadecimal' => function ($value, $options) {
+                return $this->encodeHexadecimal($value, $options);
             },
         ];
     }
@@ -67,7 +67,7 @@ class IntegerEncoder implements Encoder
      */
     public function encodeBinary($integer)
     {
-        return sprintf('%s0b%s', $this->sign($integer), decbin(abs($integer)));
+        return sprintf('%s0b%b', $this->sign($integer), abs($integer));
     }
 
     /**
@@ -77,7 +77,7 @@ class IntegerEncoder implements Encoder
      */
     public function encodeOctal($integer)
     {
-        return sprintf('%s0%s', $this->sign($integer), decoct(abs($integer)));
+        return sprintf('%s0%o', $this->sign($integer), abs($integer));
     }
 
     /**
@@ -89,7 +89,7 @@ class IntegerEncoder implements Encoder
     public function encodeDecimal($integer, $options)
     {
         if ($integer === 1 << (PHP_INT_SIZE * 8 - 1)) {
-            return sprintf('(int)%s%s', $options['whitespace'] ? ' ' : '', $integer);
+            return sprintf('(int)%s%d', $options['whitespace'] ? ' ' : '', $integer);
         }
 
         return var_export($integer, true);
@@ -98,11 +98,16 @@ class IntegerEncoder implements Encoder
     /**
      * Encodes an integer into hexadecimal representation.
      * @param int $integer The integer to encode
+     * @param array $options The integer encoding options
      * @return string The PHP code representation for the integer
      */
-    public function encodeHexadecimal($integer)
+    public function encodeHexadecimal($integer, $options)
     {
-        return sprintf('%s0x%s', $this->sign($integer), dechex(abs($integer)));
+        if ($options['hex.capitalize']) {
+            return sprintf('%s0x%X', $this->sign($integer), abs($integer));
+        }
+
+        return sprintf('%s0x%x', $this->sign($integer), abs($integer));
     }
 
     /**